fix: move database_connection fixture to manager/repositories scope
The database_connection fixture in the root conftest.py was causing
import errors across all tests because it imported from
ai.backend.manager.models.utils, which triggered a cascade of imports
that eventually tried to access manager/server.py (not available in
pants sandbox).

Moved the fixture to tests/unit/manager/repositories/conftest.py where
it properly belongs, since it's only used for repository tests.
