+++ tests/unit/manager/repositories/BUILD
@@ -4,3 +4,7 @@ python_tests(
         "src/ai/backend/manager:src",
     ],
 )
+
+python_test_utils(
+    name="test_utils",
+)
